Add LineParser
