+++ redpen-core/src/main/java/cc/redpen/parser/asciidoc/AsciiDocParser.java
@@ -18,21 +18,14 @@
 
 package cc.redpen.parser.asciidoc;
 
-import cc.redpen.RedPenException;
-import cc.redpen.model.Document;
-import cc.redpen.parser.BaseDocumentParser;
-import cc.redpen.parser.SentenceExtractor;
 import cc.redpen.parser.common.Line;
+import cc.redpen.parser.common.LineParser;
 import cc.redpen.parser.common.Model;
-import cc.redpen.tokenizer.RedPenTokenizer;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.io.BufferedReader;
 import java.io.InputStream;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Optional;
 
 
 /**
@@ -49,7 +42,7 @@ import java.util.Optional;
  * <p>
  * AsciiDoc's syntax and grammar is documented at @see http://asciidoc.org/
  */
-public class AsciiDocParser extends BaseDocumentParser {
+public class AsciiDocParser extends LineParser {
 
     private static final Logger LOG = LoggerFactory.getLogger(AsciiDocParser.class);
 
@@ -104,7 +97,7 @@ public class AsciiDocParser extends BaseDocumentParser {
     /**
      * Target line of parser
      */
-    private class TargetLine {
+    class TargetLine {
         // target line
         public Line line;
         // previous line of target line
@@ -124,23 +117,6 @@ public class AsciiDocParser extends BaseDocumentParser {
             this.firstChar = line.charAt(0);
             this.secondChar = line.charAt(1);
         }
-    }
-
-    @Override
-    public Document parse(InputStream inputStream, Optional<String> fileName, SentenceExtractor sentenceExtractor,
-                          RedPenTokenizer tokenizer) throws RedPenException {
-        Document.DocumentBuilder documentBuilder = Document.builder(tokenizer);
-        fileName.ifPresent(documentBuilder::setFileName);
-
-        Model model = new Model(sentenceExtractor);
-
-        // add the lines from the input to the model
-        populateModel(model, inputStream);
-
-        // convert the model to a redpen document
-        convertModel(model, documentBuilder);
-
-        return documentBuilder.build();
     }
 
     /**
@@ -178,69 +154,6 @@ public class AsciiDocParser extends BaseDocumentParser {
         }
         if (LOG.isDebugEnabled()) {
             LOG.debug("AsciiDoc parser model (X=erased line,[=block,section-listlevel-lineno,*=list item):\n" + model.toString());
-        }
-    }
-
-    /**
-     * Conver the parser's model to the RedPen document model
-     *
-     * @param model model to convert
-     * @param builder doc builder
-     */
-    protected void convertModel(Model model, Document.DocumentBuilder builder) {
-        model.rewind();
-
-        // add a header if there isn't one in the model
-        if ((model.getCurrentLine() != null) && (model.getCurrentLine().getSectionLevel() == 0)) {
-            builder.addSection(0);
-        }
-
-        while (model.isMore()) {
-
-            // skip blank lines
-            while (model.isMore() && model.getCurrentLine().isEmpty()) {
-                model.getNextLine();
-            }
-
-            if (model.isMore()) {
-                // check for new sections
-                if (model.getCurrentLine().getSectionLevel() > 0) {
-                    builder.addSection(
-                            model.getCurrentLine().getSectionLevel(),
-                            model.convertToSentences(model.getCurrentLine())
-                    );
-                    model.getNextLine();
-                }
-                // check for a list item
-                else if (model.getCurrentLine().isListStart()) {
-                    List<Line> listElementLines = new ArrayList<>();
-                    int listLevel = model.getCurrentLine().getListLevel();
-
-                    // add the list start line
-                    listElementLines.add(model.getCurrentLine());
-
-                    // test the following lines to see if they continue this list item
-                    model.getNextLine();
-                    while (model.isMore() &&
-                            !model.getCurrentLine().isListStart() &&
-                            (model.getCurrentLine().getListLevel() == listLevel)) {
-                        listElementLines.add(model.getCurrentLine());
-                        model.getNextLine();
-                    }
-                    builder.addListElement(listLevel, model.convertToSentences(listElementLines));
-                }
-                // process a paragraph
-                else {
-                    List<Line> paragraphLines = new ArrayList<>();
-                    // current line can't be empty, so this loop will enter at least once
-                    while (model.isMore() && !model.getCurrentLine().isEmpty()) {
-                        paragraphLines.add(model.getCurrentLine());
-                        model.getNextLine();
-                    }
-                    builder.addParagraph();
-                    model.convertToSentences(paragraphLines).forEach(builder::addSentence);
-                }
-            }
         }
     }
 
@@ -571,4 +484,3 @@ public class AsciiDocParser extends BaseDocumentParser {
         line.erase("(TM)");
     }
 }
-
